idk gradiant

+++ src/main/java/educanet/Playground.java
@@ -7,7 +7,7 @@ public class Playground {
 
     public static void main(String[] args) {
         float[] arr = {};
-        int[] arr2 = {};
+        float[] arr2 = {};
 
         Square square1 = new Square(arr, arr2, arr);
     }
